This is everything I've worked for 18/5/21.
Should be all HTML pages required.

The Javascript is not complete.
The ending page is missing javascript code to display the code name of the subject the user should enrol in and does not include links to the subject on the tafesa website or details of the lecturer to get in contact with.